vehicle/viss/0260-unsubscribeall-success.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>
[0260: end all subscription]
</title>
<link rel="help" href="https://www.w3.org/TR/vehicle-information-service/#idl-def-getrequest">
<link rel="stylesheet" type="text/css" href="./style-general.css">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="./vehicle-test-config.js"></script>
<script src="./vehicle-test-helper.js"></script>

<div id="test-title">
[0260: end all subscription]
</div>
<br>
<div id="test-assertion">
<b>Test assertion:</b><br>

A client may unsubscribe from all of its subscriptions<br>
 by sending an 'unsubscribeRequest' with the action property 'unsubscribeAll'.<br>
In this case, specifying subscriptionId value is not needed.<br>
</div>
<div id="scenario">
<br>
<b>Test Scenario:</b><br>
1. Subscribe to multiple data path.(e.g. 3 or 4)<br>
2. Confirm all subscription succeeded.<br>
3. Send 'UnsubscribeRequest' with action='unsubscribeAll'<br>
4. As the response, 'UnsubscribeSuccessResponse' will return.<br>
5. Wait for a certain time and confirm all the subscriptions are unsubscribed<br>
   and no more 'subscriptionNotification' arrive.<br>
<br>
<b>Expected Result:</b><br>
- By sending 'UnsubscribeAll' request, receives 'UnsubscribeSuccessResponse'.<br>
- After receiving 'UnsubscribeSuccessResponse', no 'SubscirbeNotification' will not reach.<br>
</div>

<div id="result"></div>
<br>
<div id="log"></div>

<script>
var reqId1 = "reqid-0261";
var reqId2 = "reqid-0262";
var reqId3 = "reqid-0263";
var subSuccess1 = false;
var subSuccess2 = false;
var subSuccess3 = false;

var reqId_unsubAll = "reqid-0264";
var subId = null;
var invalidSubId = "12345678";

var path1 = SUBSCRIBE_PATH_1;
var path2 = SUBSCRIBE_PATH_2;
var path3 = SUBSCRIBE_PATH_3;

// phase
var PH_START            = 0;
var PH_SENT_SUBSCRIBE   = 1;
var PH_SUBSCRIBED       = 2;
var PH_SENT_UNSUBSCRIBE = 3;
var PH_UNSUBSCRIBED     = 4;
var PH_END              = 5;

// Wait time to confirm SubscriptionNotification will be received.
var TIME_CONFIRM_SUB_NOTIF = 1000;
// Time to wait for all of remaining SubscriptionNotification will be received.
var TIME_WAIT_SUB_NOTIF_END = 500;
// Time to confirm all the subscription has been really unsubscribed.
// (If no more SubscriptionNotification comes, regard as unsubscribeAll succeeded.)
var TIME_CONFIRM_UNSUBSCRIBED = 1000;

var phase = PH_START;
var bNothingHasCome = true;

var t = async_test("Unsubscribe Test: unsubscribeAll test. Success case");
var vehicle  = new WebSocket(VISS_URL, VISS_SUBPROTO);

vehicle.onopen = function() {

  // process messages from the server
  vehicle.onmessage = function(event) {
    addLogMessage("Response: json="+ event.data);
    var msg = JSON.parse(event.data);

    // subscirbe has sent
    if (phase == PH_START) {
      // nothing to do
    } else if (phase == PH_SENT_SUBSCRIBE) {
      t.step_func(function() {
        // wait for subscribe response
        if (msg.action !== "subscribe") {
          addLogMessage("PH_SENT_SUBSCRIBE: Ignore messages except for Subscribe responses.");
          addLogMessage("  msg = " + JSON.stringify(msg));
          return;
        }
        if (reqId1 === msg.requestId) {
          if (isSubscribeSuccessResponse(reqId1, msg)) {
            addLogMessage("reqId1: subscribeSuccessResponse received. subId = " + msg.subscriptionId);
            subSuccess1 = true;
          }
        }
        if (reqId2 === msg.requestId) {
          if (isSubscribeSuccessResponse(reqId2, msg)) {
            addLogMessage("reqId2: subscribeSuccessResponse received. subId = " + msg.subscriptionId);
            subSuccess2 = true;
          }
        }
        if (reqId3 === msg.requestId) {
          if (isSubscribeSuccessResponse(reqId3, msg)) {
            addLogMessage("reqId3: subscribeSuccessResponse received. subId = " + msg.subscriptionId);
            subSuccess3 = true;
          }
        }
        // If all the Subscribe requests turned out successful, go to next phase
        if (subSuccess1 && subSuccess2 && subSuccess3) {
          phase = PH_SENT_UNSUBSCRIBE;

          // wait some time before execute 'unsubscribeAll' to confirm subscriptionNotification comes
          t.step_timeout(function() {

            addLogMessage("All subscribe succeeded.");
            addLogMessage("<br>Send unsubscribeAll request");
            var unsub_req = '{"action":"unsubscribeAll"'
                           +' ,"requestId":"'+reqId_unsubAll+'"}';
            addLogMessage("unsubscribeAll req: json="+unsub_req);
            vehicle.send(unsub_req);

            // Set timeout for unsubscribeAll request
            t.step_timeout(function() {
              if (phase < PH_UNSUBSCRIBED) {
                helper_terminate_failure("Finish by timeout. Response for unsubscribeAll not returned.");
              }
            }, TIME_OUT_TIME);
          }, TIME_CONFIRM_SUB_NOTIF);
        }

      })();
    } else if (phase == PH_SENT_UNSUBSCRIBE) {
      t.step_func(function() {
        if (msg.action !== "unsubscribeAll") {
          addLogMessage("PH_SENT_UNSUBSCRIBE: Ignore messages except for UnsubscribeAll response.");
          addLogMessage("  msg = " + JSON.stringify(msg));
          return;
        }
        if (isUnsubscribeAllSuccessResponse(reqId_unsubAll, msg)) {
          addLogMessage( "UnsubscribeSuccess response received." );
          // unsubscibe success
          assert_true(true, "unsubscribeAll success response received.");
          phase = PH_UNSUBSCRIBED;

          // wait for remaining subscriptionNotification exhaustion
          addLogMessage( "<br>Wait some time for remaining subscriptionNotification exhaust." );
          t.step_timeout(function() {
            // wait for some time and no subscriptionNotifcation has reached
            // judge as unsubscribeAll succeeded
            addLogMessage( "<br>Wait some time to confirm no more subscriptionNotification will come." );
            t.step_timeout(function() { // wait for confirmation time
              if (bNothingHasCome) {
                addLogMessage("Test Success: unsubscribeAll success");
                helper_terminate_success("unsubscribeAll success");
              }
            }, TIME_CONFIRM_UNSUBSCRIBED);
          }, TIME_WAIT_SUB_NOTIF_END);

        } else if (isUnsubscribeAllErrorResponse(reqId_unsubAll, "", msg)) {
          addLogMessage( "UnsubscribeAllError response received." );
          // unsubscibe error
          // Test is already success. Terminate anyway
          phase = PH_END;
          helper_terminate_failure("unsubscribeErrorResponse received");
        } else {
          addLogMessage( "Bug: This should never happen." );
        }
      })();
    } else if (phase == PH_UNSUBSCRIBED) {
      // confirm all the subscription have been unsubscribed
      // - wait for subscriptionNotification for some time
      // - if notifications continues to come, unsubscribeAll is failed
      // TODO: changed to better judgement cliteria

      t.step_func(function() {
        var bNotifSuccess = isSubscriptionNotificationResponse( "", msg);
        var bNotifError   = isSubscriptionNotificationErrorResponse( "", msg);

        if (bNotifSuccess || bNotifError) {
          // Even after received UnsubscribeAllSuccessResponse received,
          //  subscriptionNotificationResponse/ErrorResponse received.
          // Judge this as unsubscribeAll did not work effectively.
          bNothingHasCome = false;
          addLogMessage( "Some subscriptionNotification message reached event after unsubscribeAll done." );
          helper_terminate_failure("unsubscribeAll failed");
        }

      })();
    }
  }

  // Send Subscribe requests
  var sub_req_1 = '{"action":"subscribe","path":"'+ path1 +'" ,"requestId":"'+reqId1+'"}';
  addLogMessage("subscribe req1: json="+ sub_req_1);
  vehicle.send( sub_req_1 );

  var sub_req_2 = '{"action":"subscribe","path":"'+ path2 +'" ,"requestId":"'+reqId2+'"}';
  addLogMessage("subscribe req2: json="+ sub_req_2);
  vehicle.send( sub_req_2 );

  var sub_req_3 = '{"action":"subscribe","path":"'+ path3 +'" ,"requestId":"'+reqId3+'"}';
  addLogMessage("subscribe req3: json="+ sub_req_3);
  vehicle.send( sub_req_3 );

  addLogMessage("");

  phase = PH_SENT_SUBSCRIBE;

  // Set timeout for subscribe request
  t.step_timeout(function() {
    if (phase < PH_SUBSCRIBED) {
      helper_terminate_failure("Finish by timeout. Response for subscribe not returned.");
    }
  }, TIME_OUT_TIME);

}
</script>
</html>
